+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Set internal state directly when changing a routing instead of waiting for full update.
+
 ### Removed
 - Removed locking as it was already handled in underlying [hdmimatrix](https://github.com/marklynch/hdmimatrix) code